+++ geonetwork4-api/src/main/java/au/org/aodn/metadata/geonetwork/configuration/AppConstants.java
@@ -1,0 +1,6 @@
+package au.org.aodn.metadata.geonetwork.configuration;
+
+public interface AppConstants {
+    String FORMAT_ISO19115_3_2018 = "iso19115-3.2018";
+    String FORMAT_XML = "xml";
+}
