+++ geonetwork4-api/src/main/java/au/org/aodn/metadata/geonetwork/service/FIFOCache.java
@@ -1,4 +1,4 @@
-package au.org.aodn.esindexer.service;
+package au.org.aodn.metadata.geonetwork.service;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
