+++ geonetwork4-api/src/main/java/au/org/aodn/metadata/geonetwork/service/GeoNetworkService.java
@@ -1,4 +1,4 @@
-package au.org.aodn.esindexer.service;
+package au.org.aodn.metadata.geonetwork.service;
 
 import au.org.aodn.stac.model.LinkModel;
