+++ geonetwork4-api/src/main/java/au/org/aodn/metadata/geonetwork/utils/UrlUtils.java
@@ -1,4 +1,4 @@
-package au.org.aodn.esindexer.utils;
+package au.org.aodn.metadata.geonetwork.utils;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
