+++ indexer/src/test/java/au/org/aodn/esindexer/service/GeoNetworkServiceIT.java
@@ -1,13 +1,14 @@
 package au.org.aodn.esindexer.service;
 
-import au.org.aodn.cloudoptimized.model.RelationType;
+import au.org.aodn.stac.model.RelationType;
 import au.org.aodn.esindexer.BaseTestClass;
-import au.org.aodn.esindexer.configuration.AppConstants;
 import au.org.aodn.esindexer.configuration.GeoNetworkSearchTestConfig;
 
+import au.org.aodn.metadata.geonetwork.configuration.AppConstants;
 import au.org.aodn.metadata.geonetwork.exception.MetadataNotFoundException;
 import au.org.aodn.esindexer.utils.AssociatedRecordsUtil;
 import au.org.aodn.esindexer.utils.JaxbUtils;
+import au.org.aodn.metadata.geonetwork.service.GeoNetworkServiceImpl;
 import au.org.aodn.metadata.iso19115_3_2018.MDMetadataType;
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch.core.SearchRequest;
