+++ stacmodel/src/main/java/au/org/aodn/stac/model/RelationType.java
@@ -1,8 +1,7 @@
-package au.org.aodn.cloudoptimized.model;
+package au.org.aodn.stac.model;
 
 import lombok.Getter;
 
-// TODO: Should use some lib provide value or move it to stacmodel folder
 @Getter
 public enum RelationType {
     SELF("self"),
